feat(reactotron-core-contract): migrate log payload
