+++ lib/reactotron-core-contract/src/reactotron-core-contract.ts
@@ -1,2 +1,3 @@
 export * from "./command"
 export * from "./server-events"
+export * from "./log"
